requirements.txt
torch
torchvision
torchaudio
pytorch-lightning
tensorrt
onnx
onnxruntime-gpu
pycuda
numpy
timm
clip-by-openai
huggingface_hub
tqdm